refactor: update compressed_tensor for master
